senales/senales/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
autopep8==2.3.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.7.0
dateparser==1.2.0
Django==5.0
django-allauth==65.3.1
django-classy-tags==4.1.0
django-cms==4.1.4
django-countries==7.6.1
django-entangled==0.5.4
django-extensions==3.2.3
django-filer==3.3.1
django-formtools==2.5.1
django-fsm==2.8.2
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==4.0.0
djangocms-bootstrap5==0.1.0
djangocms-frontend==1.3.5
djangocms-icon==2.1.0
djangocms-link==5.0.0
djangocms-picture==4.1.1
djangocms-text==0.4.0
djangocms-text-ckeditor==5.1.6
djangocms-versioning==2.1.0
easy-thumbnails==2.10
frozenlist==1.5.0
html5lib==1.1
idna==3.10
lxml==5.3.0
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nh3==0.2.20
packaging==24.2
pathspec==0.12.1
paypalrestsdk==1.13.3
pillow==11.0.0
pipdeptree==2.24.0
platformdirs==4.3.6
propcache==0.2.1
psycopg==3.2.3
psycopg2==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pyOpenSSL==24.3.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
